Reduce files
