+++ tsconfig.build.json
@@ -1,9 +1,11 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    "diagnostics": true,
     "lib": ["es5"],
     "target": "es5",
-    "noEmit": false
+    "noEmit": false,
+    "noEmitOnError": true
   },
   "files": ["./types/index.d.ts", "./lib/index.ts"],
   "include": ["lib/**/*.ts", "register.ts"],
